+++ services/psibase_tests/CMakeLists.txt
@@ -16,7 +16,7 @@ function(add suffix)
     service(EmitEvents "${suffix}" src/EmitEvents.cpp)
     service(TestMemo "${suffix}" src/TestMemo.cpp)
     service(TestClock "${suffix}" src/TestClock.cpp)
-    service(subjective-service "${suffix}" subjective-service.cpp)
+    service(SubjectiveDb "${suffix}" src/SubjectiveDb.cpp)
     service(RemoveCode "${suffix}" RemoveCode.cpp)
     service(AbortService "${suffix}" AbortService.cpp)
     service(EntryPoint "${suffix}" src/EntryPoint.cpp)
